属性，rpc修改

+++ common/util/src/main/java/com/whk/message/gamegate/ReqPlayerListMessage.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 public class ReqPlayerListMessage extends ReqMessage {
     private long userId;
-    private int serverId;
+    private long serverId;
 }
