+++ game/game-server/src/main/java/com/whk/net/RpcGameProxyHolder.java
@@ -28,12 +28,12 @@ public class RpcGameProxyHolder extends RpcServerProxy {
     }
 
     @Override
-    public Server getServer(int serverId) {
+    public Server getServer(long serverId) {
         return GameServerManager.getInstance().getServer(serverId);
     }
 
     @Override
-    public String rpcRequestTopic(int serverId) {
+    public String rpcRequestTopic(long serverId) {
         return gameServerConfig.getRpcRequestTopic(serverId);
     }
